# contracts/openapi/v1/secondbox.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "SecondBox API",
    "version": "1.0.0",
    "description": "Provider-neutral administration and durable Sandbox API."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "serviceAccountBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/sandboxes/{sandboxId}/snapshots": {
      "get": {
        "tags": ["Snapshots"],
        "operationId": "listSandboxSnapshots",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/PageLimit"},
          {"$ref": "#/components/parameters/PageCursor"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "200": {
            "description": "Retained Snapshot page.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotPage"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["Snapshots"],
        "operationId": "createSandboxSnapshot",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/IfMatch"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateSnapshotRequest"}}}
        },
        "responses": {
          "201": {
            "description": "Immutable projection of the current published stopped-state checkpoint.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Snapshot"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "412": {"$ref": "#/components/responses/PreconditionFailed"},
          "429": {"$ref": "#/components/responses/TooManyRequests"}
        }
      }
    },
    "/v1/snapshots/{snapshotId}": {
      "get": {
        "tags": ["Snapshots"],
        "operationId": "getSnapshot",
        "parameters": [
          {"$ref": "#/components/parameters/SnapshotID"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "200": {
            "description": "Retained Snapshot metadata.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Snapshot"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "delete": {
        "tags": ["Snapshots"],
        "operationId": "deleteSnapshot",
        "parameters": [
          {"$ref": "#/components/parameters/SnapshotID"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "204": {
            "description": "Snapshot retention ended.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"}
        }
      }
    },
    "/v1/sandboxes/{sandboxId}/port-sessions": {
      "post": {
        "tags": ["Ports"],
        "operationId": "createSandboxPortSession",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/Generation"},
          {"$ref": "#/components/parameters/LeaseIDHeader"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreatePortSessionRequest"}}}
        },
        "responses": {
          "201": {
            "description": "Authenticated expiring port proxy session.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PortSession"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "429": {"$ref": "#/components/responses/TooManyRequests"}
        }
      }
    },
    "/v1/sandboxes/{sandboxId}/port-sessions/{portSessionId}": {
      "get": {
        "tags": ["Ports"],
        "operationId": "getSandboxPortSession",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/PortSessionID"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "200": {
            "description": "Port session.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PortSession"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "delete": {
        "tags": ["Ports"],
        "operationId": "closeSandboxPortSession",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/PortSessionID"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "204": {
            "description": "Port session closed.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "serviceAccountBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SecondBox service-account API key"
      }
    },
    "headers": {
      "RequestID": {
        "description": "End-to-end request correlation identifier.",
        "schema": {"$ref": "#/components/schemas/CorrelationID"}
      },
...
    },
    "parameters": {
      "RequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "schema": {"$ref": "#/components/schemas/CorrelationID"}
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 8, "maxLength": 200, "pattern": "^[A-Za-z0-9._~:+/=-]+$"}
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 3, "maxLength": 160}
      },
      "Generation": {
        "name": "SecondBox-Generation",
        "in": "header",
        "required": true,
        "schema": {"type": "integer", "format": "int64", "minimum": 1}
      },
      "LeaseIDHeader": {
        "name": "SecondBox-Lease-ID",
        "in": "header",
        "required": false,
        "schema": {"$ref": "#/components/schemas/OpaqueID"}
      },
...
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {"type": "integer", "minimum": 1, "maximum": 200}
      },
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {"type": "string", "minLength": 1, "maxLength": 512}
      },
...
      "SandboxID": {"name": "sandboxId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/OpaqueID"}},
...
      "SnapshotID": {"name": "snapshotId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/OpaqueID"}},
      "PortSessionID": {"name": "portSessionId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/OpaqueID"}},
...
    },
...
    "responses": {
...
      "BadRequest": {"$ref": "#/components/responses/Problem400"},
      "Unauthorized": {"$ref": "#/components/responses/Problem401"},
      "Forbidden": {"$ref": "#/components/responses/Problem403"},
      "NotFound": {"$ref": "#/components/responses/Problem404"},
...
      "Conflict": {"$ref": "#/components/responses/Problem409"},
      "PreconditionFailed": {"$ref": "#/components/responses/Problem412"},
...
      "TooManyRequests": {"$ref": "#/components/responses/Problem429"},
      "Problem400": {
        "description": "Invalid request.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem401": {
        "description": "Authentication failed.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem403": {
        "description": "Authorization failed.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem404": {
        "description": "Resource not found in the authenticated scope.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
...
      "Problem409": {
        "description": "Current state, generation, Lease, quota, or idempotency conflict.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem412": {
        "description": "Resource revision does not match If-Match.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
...
      "Problem429": {
        "description": "Quota or concurrent-operation limit exceeded.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      }
    },
    "schemas": {
      "OpaqueID": {"type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9_-]+$"},
      "CorrelationID": {"type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9._:-]+$"},
      "Timestamp": {"type": "string", "format": "date-time"},
...
      "Metadata": {
        "type": "object",
        "maxProperties": 32,
        "propertyNames": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9._/-]+$"},
        "additionalProperties": {"type": "string", "maxLength": 1024}
      },
...
      "Snapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sandboxId", "generation", "name", "sizeBytes", "sha256", "metadata", "expiresAt", "createdAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "sandboxId": {"$ref": "#/components/schemas/OpaqueID"},
          "generation": {"type": "integer", "format": "int64", "minimum": 1},
          "name": {"type": "string", "minLength": 1, "maxLength": 255},
          "sizeBytes": {"type": "integer", "format": "int64", "minimum": 0},
          "sha256": {"type": "string", "pattern": "^[a-f0-9]{64}$"},
          "metadata": {"$ref": "#/components/schemas/Metadata"},
          "expiresAt": {"$ref": "#/components/schemas/Timestamp"},
          "createdAt": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
      "CreateSnapshotRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "metadata"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 255},
          "metadata": {"$ref": "#/components/schemas/Metadata"}
        }
      },
      "SnapshotPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items"],
        "properties": {
          "items": {"type": "array", "maxItems": 200, "items": {"$ref": "#/components/schemas/Snapshot"}},
          "nextCursor": {"type": "string", "minLength": 1, "maxLength": 512}
        }
      },
      "CreatePortSessionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "durationSeconds"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 80},
          "durationSeconds": {"type": "integer", "minimum": 1, "maximum": 86400}
        }
      },
      "PortSession": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sandboxId", "generation", "name", "protocol", "endpoint", "state", "createdAt", "expiresAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "sandboxId": {"$ref": "#/components/schemas/OpaqueID"},
          "generation": {"type": "integer", "format": "int64", "minimum": 1},
          "name": {"type": "string", "minLength": 1, "maxLength": 80},
          "protocol": {"type": "string", "enum": ["tcp", "http"]},
          "endpoint": {"type": "string", "format": "uri", "maxLength": 2048},
          "state": {"type": "string", "enum": ["open", "closing", "closed", "expired", "fenced"]},
          "createdAt": {"$ref": "#/components/schemas/Timestamp"},
          "expiresAt": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
      "ProblemCode": {
        "type": "string",
        "enum": [
          "invalid_request",
          "authentication_failed",
          "authorization_failed",
          "not_found",
          "idempotency_conflict",
          "precondition_failed",
          "state_conflict",
          "generation_fenced",
          "lease_fenced",
          "profile_unavailable",
          "quota_exceeded",
          "limit_exceeded",
          "guest_unavailable",
          "execution_node_unavailable",
          "dependency_unavailable",
          "internal_error",
          "wait_expired"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["field", "reason"],
        "properties": {
          "field": {"type": "string", "minLength": 1, "maxLength": 256},
          "reason": {"type": "string", "minLength": 1, "maxLength": 1024}
        }
      },
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "code", "requestId", "retryable"],
        "properties": {
          "type": {"type": "string", "format": "uri", "maxLength": 2048},
          "title": {"type": "string", "minLength": 1, "maxLength": 256},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "code": {"$ref": "#/components/schemas/ProblemCode"},
          "requestId": {"$ref": "#/components/schemas/CorrelationID"},
          "retryable": {"type": "boolean"},
          "details": {"type": "array", "maxItems": 32, "items": {"$ref": "#/components/schemas/ProblemDetail"}}
        }
      }
    }
  }
}